# Fixed #28948 -- Optimized CookieStorage performances.