@David-Wobrock 2023-03-26
> I would also be very interested to see (if it is not to much work) a comparison of the old & new code with a single message (Current code vs this PR) -- while it is good that we improve the case for many messages we should also ensure that we do not regress in the common case where the user is not malicious. The code doesn't seem to suggest that we are doing so on a big scale, but given that you probably already have the setup to check this it would still be nice to see.

Good point 👍 

Intuitively, it shouldn't change much because the amount of computation done is the same when the messages doesn't exceed the max cookie size.
Just, instead of calling one function that encodes messages + compress/signs them, we split this into 2 different function calls (more or less).

<details><summary>The Python timing code</summary>
<p>

```python
from django.conf import settings
settings.configure(MESSAGE_TAGS={}, SECRET_KEY="x")
from django.contrib.messages.storage.cookie import CookieStorage
from django.http.request import HttpRequest
from django.http.response import HttpResponse
from timeit import timeit

storage = CookieStorage(HttpRequest())
storage.add(20, f"amessage")
%timeit storage.update(HttpResponse())
```

</p>
</details>


On `main`:
```
20.9 µs ± 13 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

On this branch:
```
21.4 µs ± 20.4 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```