@apollo13 2023-03-26
> Intuitively, it shouldn't change much because the amount of computation done is the same when the messages doesn't exceed the max cookie size.

Yeah, that was my thought after reading the code as well. The timings for a single message look okay :+1: